requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
defusedxml==0.7.1
django-bootstrap-icons==0.9.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-feather==0.3.0
django-formset==2.1
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.1
phonenumbers==9.0.12
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.1
service-identity==24.2.0
setuptools==80.9.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
zope.interface==7.2
